[Testing] Fixed Type When Reading Lookahead File
There was a typo when reading in the router lookahead file where it was
checking for the existence of the rr_graph file instead of the router
lookahead file. This may cause a crash when the rr_graph file is
provided, but the router lookahead file is not. It also would not
function properly if the rr_graph file is not provided.

Thank you to @nedsels for pointing this out.